Moved deferred pool, connection and executor to respectful files
