+++ .gitignore
@@ -126,3 +126,5 @@ lib
 e2e/esmCompatibility/.output
 src/e2e/esmCompatibility/.output
 **/0x
+
+**/*.db
