+++ src/packages/dumbo/src/core/connections/connection.ts
@@ -1,5 +1,6 @@
 import type { ConnectorType } from '../connectors';
 import {
+  createDeferredExecutor,
   sqlExecutor,
   type DbSQLExecutor,
   type WithSQLExecutor,
@@ -84,4 +85,54 @@ export const createConnection = <
   const typedConnection = connection as ConnectionType;
 
   return typedConnection;
+};
+
+export const createDeferredConnection = <Connector extends ConnectorType>(
+  connector: Connector,
+  importConnection: () => Promise<Connection<Connector>>,
+): Connection<Connector> => {
+  const getConnection = importConnection();
+
+  const execute = createDeferredExecutor(async () => {
+    const conn = await getConnection;
+    return conn.execute;
+  });
+
+  const connection: Connection<Connector> = {
+    connector,
+    execute,
+
+    open: async (): Promise<unknown> => {
+      const conn = await getConnection;
+      return conn.open();
+    },
+
+    close: async (): Promise<void> => {
+      if (getConnection) {
+        const conn = await getConnection;
+        await conn.close();
+      }
+    },
+
+    transaction: () => {
+      const transaction = getConnection.then((c) => c.transaction());
+
+      return {
+        connector,
+        connection,
+        execute: createDeferredExecutor(
+          async () => (await transaction).execute,
+        ),
+        begin: async () => (await transaction).begin(),
+        commit: async () => (await transaction).commit(),
+        rollback: async () => (await transaction).rollback(),
+      };
+    },
+    withTransaction: async (handle) => {
+      const connection = await getConnection;
+      return connection.withTransaction(handle);
+    },
+  };
+
+  return connection;
 };
