+++ src/packages/dumbo/src/index.ts
@@ -1,1 +1,2 @@
 export * from './core';
+export * from './storage/all';
